// frontend/src/api-client/models/McpToolDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type McpToolDto = {
    id: string;
    name: string;
    description: string;
    inputSchema: Record<string, any>;
    /**
     * The ID of the MCP server this tool belongs to
     */
    serverId: string;
};